+++ src/game/Tactical/PathAI.cc
@@ -2312,7 +2312,7 @@ INT16 PlotPath(SOLDIERTYPE* const pSold, const INT16 sDestGridno, const INT8 bCo
 
 			//if (gTacticalStatus.uiFlags & INCOMBAT) // OR USER OPTION "show paths" ON... ***
 			{
-				if (bPlot && iCnt < iLastGrid - 1 && giPlotCnt < lengthof(guiPlottedPath))
+				if (bPlot && iCnt < iLastGrid - 1 && giPlotCnt < static_cast<INT32>(lengthof(guiPlottedPath)))
 				{
 					guiPlottedPath[giPlotCnt++] = sTempGrid;
